prune data_for_model_fit.json
