+++ pipelines/forecast_timeseries.py
@@ -242,7 +242,6 @@ def main(
         report_date=report_date,
         first_training_date=first_training_date,
         last_training_date=last_training_date,
-        param_estimates=param_estimates,
         model_run_dir=model_run_dir,
         logger=logger,
         credentials_dict=credentials_dict,
